api_bad/requirements.txt
flask
flask-sqlalchemy
onetimepad
pycrypto
boto3
pymysql
flask-script
flask_cors
serializer